deployment/examples/bentoml_deployment_patterns/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --no-hashes
a2wsgi==1.10.8
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
attrs==25.1.0
bentoml==1.3.20
cattrs==23.1.2
certifi==2025.1.31
cffi==1.17.1 ; implementation_name == 'pypy'
charset-normalizer==3.4.1
click==8.1.8
click-option-group==0.5.6
cloudpickle==3.1.1
colorama==0.4.6 ; sys_platform == 'win32'
coloredlogs==15.0.1
deepmerge==2.0
deprecated==1.2.18
filelock==3.17.0
flatbuffers==25.1.24
frozenlist==1.5.0
fs==2.4.16
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-ws==0.7.1
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
importlib-metadata==8.5.0
inflection==0.5.1
jinja2==3.1.5
kantoku==0.18.1
light-embed==1.0.7
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
numpy==1.26.4
nvidia-ml-py==12.570.86
onnxruntime==1.19.2
opentelemetry-api==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-aiohttp-client==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
packaging==24.2
pathspec==0.12.1
pip-requirements-parser==32.0.1
prometheus-client==0.21.1
prompt-toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pycparser==2.22 ; implementation_name == 'pypy'
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pyparsing==3.2.1
pyreadline3==3.5.4 ; sys_platform == 'win32'
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pyyaml==6.0.2
pyzmq==26.2.1
questionary==2.1.0
requests==2.32.3
rich==13.9.4
schema==0.7.7
setuptools==75.8.0
simple-di==0.1.5
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.3
tokenizers==0.20.1
tomli-w==1.2.0
tornado==6.4.2
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.3.0
uv==0.5.26
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
